# Use vips reduce when downscaling