I'm scaling an image down by about half and it seems bimg is using vips_affine to do the downscaling, which results in really aliased images:
![image](https://cloud.githubusercontent.com/assets/2247982/24089913/bb9dc6e6-0d8f-11e7-9d07-11792359a24b.png)


Instead I've swapped it to use vips_reduce when downscaling:
![image](https://cloud.githubusercontent.com/assets/2247982/24089881/5f2efb1e-0d8f-11e7-98ec-9f4f45439d47.png)

I think this would fix https://github.com/h2non/bimg/issues/94