@Dynom 2017-05-25
That is a significant improvement! What about performance, is there a significant difference between the two approaches?