@h2non 2017-05-25
Adding some tests here would be nice.