@h2non 2017-05-25
@Vektah can you please fix this formatting inconsistency?
https://travis-ci.org/h2non/bimg/jobs/212880888#L2919-L2927