@Vektah 2017-07-04
Baseline
```
roci :: 99designs/bimg » go test -bench=. .
BenchmarkRotateJpeg-8                         10         101282015 ns/op
BenchmarkResizeLargeJpeg-8                    30          48038841 ns/op
BenchmarkResizePng-8                         100          36754256 ns/op
BenchmarkResizeWebP-8                         50          22225485 ns/op
BenchmarkConvertToJpeg-8                     100          17945110 ns/op
BenchmarkConvertToPng-8                       10         157521422 ns/op
BenchmarkConvertToWebp-8                      10         160794332 ns/op
BenchmarkCropJpeg-8                           30          48285443 ns/op
BenchmarkCropPng-8                            10         122325220 ns/op
BenchmarkCropWebP-8                           20          93161352 ns/op
BenchmarkExtractJpeg-8                       100          21483165 ns/op
BenchmarkExtractPng-8                       2000            876224 ns/op
BenchmarkExtractWebp-8                     10000            194227 ns/op
BenchmarkZoomJpeg-8                           20          96650448 ns/op
BenchmarkZoomPng-8                            30          67097133 ns/op
BenchmarkZoomWebp-8                           20          97779698 ns/op
BenchmarkWatermarkJpeg-8                      30          51219012 ns/op
BenchmarkWatermarPng-8                      1000           1721124 ns/op
BenchmarkWatermarWebp-8                       30          41610667 ns/op
BenchmarkWatermarkImageJpeg-8                 20          79063884 ns/op
BenchmarkWatermarImagePng-8                   30          59174326 ns/op
BenchmarkWatermarImageWebp-8                  20          68209704 ns/op
PASS
ok      github.com/99designs/bimg       48.070s
```
With changes:
```
roci :: 99designs/bimg » go test -bench=. .
BenchmarkRotateJpeg-8                         10         104976036 ns/op
BenchmarkResizeLargeJpeg-8                    30          50475049 ns/op
BenchmarkResizePng-8                          50          39329219 ns/op
BenchmarkResizeWebP-8                        100          23175726 ns/op
BenchmarkConvertToJpeg-8                     100          17689629 ns/op
BenchmarkConvertToPng-8                       10         154330459 ns/op
BenchmarkConvertToWebp-8                      10         162643603 ns/op
BenchmarkCropJpeg-8                           30          51620126 ns/op
BenchmarkCropPng-8                            10         121402841 ns/op
BenchmarkCropWebP-8                           20          95114489 ns/op
BenchmarkExtractJpeg-8                       100          21732163 ns/op
BenchmarkExtractPng-8                       2000            874188 ns/op
BenchmarkExtractWebp-8                     10000            193840 ns/op
BenchmarkZoomJpeg-8                           20          93977438 ns/op
BenchmarkZoomPng-8                            30          69719061 ns/op
BenchmarkZoomWebp-8                           20          98654028 ns/op
BenchmarkWatermarkJpeg-8                      30          51335162 ns/op
BenchmarkWatermarPng-8                      1000           1704559 ns/op
BenchmarkWatermarWebp-8                       30          40709982 ns/op
BenchmarkWatermarkImageJpeg-8                 20          80133307 ns/op
BenchmarkWatermarImagePng-8                   30          58825332 ns/op
BenchmarkWatermarImageWebp-8                  20          67748778 ns/op
PASS
ok      github.com/99designs/bimg       48.905s
```

Looks to be a little variation there so I did a longer benchmark on the large resize:
```
before          500          49679403 ns/op
after           500          52067031 ns/op
```

Looks to be ~4% slower?

How do you test image quality changes? I think the codepath itself is already pretty well covered by resize_test.go