@Dynom 2017-07-10
Thanks @Vektah. 4% would be an acceptable trade-off for me for these benchmarks.