@rgritti 2017-08-16
I am looking forward for this PR to be merged. Can somebody merge it?