@h2non 2017-09-10
Feature available in `bimg@1.0.12`.