@martijnotto 2017-11-13
This PR means that bimg no longer compiles with libvips 8.2, since this does not have vips_reduce. This now means the README.md is no longer correct, since it only _recommends_ 8.4+

Perhaps it's an idea to add some preprocessor statements to fall back to using vips_affine in this case?